适配psr-4 autoload及整理代码
