+++ _app/lib/Abstract.php
