+++ _app/lib/Errors/NoPermission.php
@@ -1,0 +1,7 @@
+<?php
+namespace xyToki\xyShare\Errors;
+class NoPermission implements \Throwable{
+    public function getMessage(){
+        return "无权限";
+    }
+}
