+++ _app/lib/Errors/NotAuthorized.php
@@ -1,0 +1,7 @@
+<?php
+namespace xyToki\xyShare\Errors;
+class NotAuthorized implements \Throwable{
+    public function getMessage(){
+        return "未授权";
+    }
+}
