+++ _app/lib/Errors/NotConfigured.php
@@ -1,0 +1,7 @@
+<?php
+namespace xyToki\xyShare\Errors;
+class NotConfigured implements \Throwable{
+    public function getMessage(){
+        return "配置不正确";
+    }
+}
