+++ _app/lib/Errors/NotFound.php
@@ -1,0 +1,7 @@
+<?php
+namespace xyToki\xyShare\Errors;
+class NotFound implements \Throwable{
+    public function getMessage(){
+        return "文件不存在";
+    }
+}
