--- _app/lib/errors.class.php
@@ -1,14 +1,0 @@
-<?php
-namespace xyToki\xyShare\Errors;
-class NotAuthorized extends \Error{
-
-}
-class NotConfigured extends \Error{
-
-}
-class NotFound extends \Error{
-
-}
-class NoPermission extends \Error{
-
-}
